+++ x/wasm/internal/types/genesis_test.go
@@ -127,12 +127,6 @@ func TestContractValidateBasic(t *testing.T) {
 			},
 			expError: true,
 		},
-		"contract with history set": {
-			srcMutator: func(c *Contract) {
-				c.ContractInfo.ContractCodeHistory = []ContractCodeHistoryEntry{{}}
-			},
-			expError: true,
-		},
 		"contract state invalid": {
 			srcMutator: func(c *Contract) {
 				c.ContractState = append(c.ContractState, Model{})