+++ x/wasm/internal/types/keys.go
@@ -31,12 +31,14 @@ const ( // event attributes
 
 // nolint
 var (
-	KeyLastCodeID     = []byte("lastCodeId")
-	KeyLastInstanceID = []byte("lastContractId")
+	CodeKeyPrefix              = []byte{0x01}
+	ContractKeyPrefix          = []byte{0x02}
+	ContractStorePrefix        = []byte{0x03}
+	SequenceKeyPrefix          = []byte{0x04}
+	ContractHistoryStorePrefix = []byte{0x05}
 
-	CodeKeyPrefix       = []byte{0x01}
-	ContractKeyPrefix   = []byte{0x02}
-	ContractStorePrefix = []byte{0x03}
+	KeyLastCodeID     = append(SequenceKeyPrefix, []byte("lastCodeId")...)
+	KeyLastInstanceID = append(SequenceKeyPrefix, []byte("lastContractId")...)
 )
 
 // GetCodeKey constructs the key for retreiving the ID for the WASM code